requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.0.1
Django==4.2
django-cloudinary-storage==0.3.0
django-environ==0.12.0
django-widget-tweaks==1.5.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
